DBZ-7192: move abandoned transactions impl to Abstract layer
So it is available for ISPN. ISPN does not store list of abandoned transactions

fix rebase

cosmetic changes
